test: move oss bucket test into shared
